requirements.txt
pytest==4.6.0
selenium==3.141.0
pillow==6.1.0
pysqlite3=0.4.1